requirements.txt
playwright
python-dotenv
gspread
oauth2client
google-generativeai
requests
